+++ clarifai/runners/__init__.py
@@ -1,11 +1,11 @@
 from .models.base_typed_model import AnyAnyModel, TextInputModel, VisualInputModel
 from .models.model_runner import ModelRunner
-from .models.model_upload import ModelUploader
+from .models.model_builder import ModelBuilder
 from .utils.data_handler import InputDataHandler, OutputDataHandler
 
 __all__ = [
     "ModelRunner",
-    "ModelUploader",
+    "ModelBuilder",
     "InputDataHandler",
     "OutputDataHandler",
     "AnyAnyModel",
